requirements.txt
numpy==1.21.0
opencv-python==4.5.1.48
Pillow==8.1.2
redis==3.5.3